requirements.txt
appnope==0.1.3
asttokens==2.0.5
autopep8==1.6.0
backcall==0.2.0
blinker==1.5
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
decorator==5.1.1
executing==0.9.1
Flask==2.2.1
Flask-DebugToolbar==0.13.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
idna==3.3
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.30
psycopg2==2.9.3
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.9.1
Pygments==2.12.0
python-dotenv==0.20.0
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.39
stack-data==0.3.0
toml==0.10.2
traitlets==5.3.0
urllib3==1.26.11
wcwidth==0.2.5
Werkzeug==2.2.1
WTForms==3.0.1
